GD-74: Improve the CI workflow for job `unit-test-example` by validate the test report
# Why
The current job unit-test-example executes the test and forces the step to always end with success.
We have done this so as not to interrupt the job.
But we forgot a validation step right after the test results that checks the expected test results.

# What
- fixed actions to publish and upload test reports to include the examples result.
- add validation step to verify for expected test results
