+++ .github/actions/publish-test-report/action.yml
@@ -1,19 +1,22 @@
 name: publish-test-report
-description: "Publishes the GdUnit test results"
+description: 'Publishes the GdUnit test results'
 
 inputs:
   report-name:
-    description: "Name of the check run which will be created."
+    description: 'Name of the check run which will be created.'
+    required: true
+  report-path:
+    description: 'The path to the reports'
     required: true
 
 runs:
   using: composite
   steps:
-    - name: "Publish Test Results"
+    - name: 'Publish Test Results'
       uses: dorny/test-reporter@v1.8.0
       with:
         name: ${{ inputs.report-name }}
-        path: "./test/TestResults/test-result.trx"
+        path: ${{ inputs.report-path }}
         reporter: dotnet-trx
-        fail-on-error: "false"
-        fail-on-empty: "false"
+        fail-on-error: 'false'
+        fail-on-empty: 'false'
