+++ .github/workflows/unit-tests.yml
@@ -39,7 +39,7 @@ jobs:
       actions: write
       contents: read
 
-    name: "Unit Tests With GdUnit4 TestRunner"
+    name: "Run gdUnit4-action"
     runs-on: ${{ inputs.os }}
     timeout-minutes: 5
 
@@ -58,15 +58,15 @@ jobs:
             res://test/src
           timeout: 5
           retries: 3 # We have set the number of repetitions to 3 because Godot mono randomly crashes during C# tests
-          report-name: report_Godot${{ inputs.godot-version }}-${{ inputs.godot-status }}-net.xml
+          report-name: Report_GdUnit4Action-${{ inputs.godot-version }}-${{ inputs.godot-status }}
 
 
   unit-test-adapter:
-    name: "Unit Tests With VS-TestAdapter"
+    name: "Run API Tests"
     runs-on: ${{ inputs.os }}
     timeout-minutes: 5
     env:
-      REPORT_NAME: "TestAdapterReport_${{ inputs.os }}-${{ inputs.godot-version }}"
+      REPORT_NAME: "Report_API_Tests-${{ inputs.godot-version }}-${{ inputs.godot-status }}"
 
     steps:
       - uses: actions/checkout@v4
@@ -109,6 +109,7 @@ jobs:
         uses: ./.github/actions/publish-test-report
         with:
           report-name: ${{ env.REPORT_NAME }}
+          report-path: './test/TestResults/test-result.trx'
 
       - name: "Upload Unit Test Reports"
         if: ${{ !cancelled() }}
@@ -117,11 +118,11 @@ jobs:
           report-name: ${{ env.REPORT_NAME }}
 
   unit-test-example:
-    name: "Unit Tests Example project using nuget gdUnit4.test.adapter"
+    name: "Run Test Examples"
     runs-on: ${{ inputs.os }}
     timeout-minutes: 3
     env:
-      REPORT_NAME: "Example_Report_${{ inputs.os }}-${{ inputs.godot-version }}"
+      REPORT_NAME: "Report_Examples-${{ inputs.godot-version }}-${{ inputs.godot-status }}"
 
     steps:
       - uses: actions/checkout@v4
@@ -157,13 +158,42 @@ jobs:
           GODOT_BIN: "/home/runner/godot-linux/godot"
         run: |
           $GODOT_BIN --path ./example --headless --build-solutions --quit-after 20
-          xvfb-run --auto-servernum dotnet test ./example/exampleProject.csproj --no-build --settings ./test/.runsettings-ci || true
+          xvfb-run --auto-servernum dotnet test ./example/exampleProject.csproj --no-build --settings ./example/.runsettings-ci || true
+
+        # Step to validate test results
+      - name: Validate Test Results
+        shell: bash
+        run: |
+          validate_section() {
+            local section="$1"
+            local expected_content="$2"
+
+            grep -A 10 "$section" ./example/TestResults/test-result.trx | sed -e 's/^[ \t]*//' | tee output.txt >/dev/null
+            if grep -qF "$expected_content" output.txt; then
+              echo "Test validation passed. The expected text is present in $section."
+            else
+              echo "Test validation failed. The " + $expected_content + " is not present in:"
+              echo ""
+              cat output.txt
+              echo ""
+              exit 1
+            fi
+          }
+
+          validate_section '<Output>' 'Discover tests done, 1 TestSuites and total 2 Tests found.'
+          validate_section '<Output>' 'Start executing tests, 2 TestCases total.'
+          validate_section '<Results>' "<Message>Expecting: 'True' but is 'False'</Message>"
+          validate_section '<Results>' '<StackTrace>   at Examples.ExampleTest.failed() in /home/runner/work/gdUnit4Mono/gdUnit4Mono/example/test/ExampleTest.cs:line 19'
+          validate_section '<ResultSummary' '<ResultSummary outcome="Failed">'
+          validate_section '<ResultSummary' '<Counters total="2" executed="2" passed="1" failed="1" error="0" timeout="0" aborted="0" inconclusive="0" passedButRunAborted="0" notRunnable="0" notExecuted="0" disconnected="0" warning="0" completed="0" inProgress="0" pending="0" />'
+          echo "All tests passes."
 
       - name: "Publish Unit Test Reports"
         if: ${{ !cancelled() }}
         uses: ./.github/actions/publish-test-report
         with:
           report-name: ${{ env.REPORT_NAME }}
+          report-path: './example/TestResults/test-result.trx'
 
       - name: "Upload Unit Test Reports"
         if: ${{ !cancelled() }}
