+++ .vscode/settings.json
@@ -26,6 +26,7 @@
     "Clazz",
     "Cmdline",
     "configfile",
+    "dorny",
     "Extr",
     "Fullqualified",
     "Fuzzer",
