+++ clean.sh
@@ -1,7 +1,6 @@
 rm -rf ./bin
 rm -rf ./obj
-rm -rf ./api/nupkg/net7.0/
-rm -rf ./api/nupkg/*.nupkg
+rm -rf ./api/nupkg/*
 rm -rf ./api/obj
 
 rm -rf .godot
@@ -13,8 +12,7 @@ rm -rf ./example/.godot
 rm -rf ./example/gdunit4_testadapter
 
 
-rm -rf ./testadapter/nupkg/net7.0/
-rm -rf ./testadapter/nupkg/*.nupkg
+rm -rf ./testadapter/nupkg/*
 rm -rf ./testadapter/obj
 
 
@@ -23,6 +21,8 @@ dotnet build ./testadapter/gdUnit4TestAdapter.csproj
 dotnet restore
 dotnet build
 
+
+$GODOT_BIN --path ./example --headless --build-solutions --quit-after 20
 # dotnet clean
 # dotnet restore
 # dotnet build
