+++ example/.runsettings
@@ -3,7 +3,7 @@
     <RunConfiguration>
         <MaxCpuCount>1</MaxCpuCount>
         <ResultsDirectory>./TestResults</ResultsDirectory>
-        <TargetFrameworkVersion>net7.0</TargetFrameworkVersion>
+        <TargetFrameworks>net7.0;net8.0</TargetFrameworks>
         <TestSessionTimeout>180000</TestSessionTimeout>
         <TreatNoTestsAsError>true</TreatNoTestsAsError>
     </RunConfiguration>
@@ -33,6 +33,6 @@
         <Parameters></Parameters>
         <!-- Controlls the Display name attribute of the TestCase. Allowed values are SimpleName and FullyQualifiedName.
              This likely determines how the test names are displayed in the test results.-->
-        <DisplayName>FullyQualifiedName</DisplayName>
+        <DisplayName>SimpleName</DisplayName>
     </GdUnit4>
 </RunSettings>
