+++ example/.runsettings-ci
@@ -28,27 +28,10 @@
         </Loggers>
     </LoggerRunSettings>
 
-    <DataCollectionRunSettings>
-        <DataCollectors>
-        <!-- Enables blame -->
-        <DataCollector friendlyName="blame" enabled="True">
-            <Configuration>
-            <!-- Enables crash dump, with dump type "Full" or "Mini".
-            Requires ProcDump in PATH for .NET Framework. -->
-            <CollectDump DumpType="Full" />
-            <!-- Enables hang dump or testhost and its child processes
-            when a test hangs for more than 10 minutes.
-            Dump type "Full", "Mini" or "None" (just kill the processes). -->
-            <CollectDumpOnTestSessionHang TestTimeout="10min" HangDumpType="Full" />
-            </Configuration>
-        </DataCollector>
-        </DataCollectors>
-    </DataCollectionRunSettings>
-
     <GdUnit4>
         <!-- Additonal Godot runtime parameters-->
         <!-- These parameters are crucial for configuring the Godot runtime to work in headless environments, such as those used in automated testing or CI/CD pipelines.-->
-        <Parameters>--audio-driver Dummy --display-driver x11 --rendering-driver opengl3 --screen 0 --verbose</Parameters>
+        <Parameters>--audio-driver Dummy --display-driver x11 --rendering-driver opengl3 --screen 0</Parameters>
         <!-- Controlls the Display name attribute of the TestCase. Allowed values are SimpleName and FullyQualifiedName.
              This likely determines how the test names are displayed in the test results.-->
         <DisplayName>SimpleName</DisplayName>
